refactor(web): split Group 5 (final) routes into APIRouters, completing Phase 6b
Moves the last domain routes still inline in create_app() — search, panel/
overview/charts/config/mapillary/serious/face, brain graph view, saude,
vault, local-agent, kb, budgets, goals, notify, plus the leftover chat/
threads/history/cmd/commands/briefing/greeting surface — into
ev/interfaces/web/routes/*.py APIRouters. app.py now only instantiates
WebContext, registers routers, and keeps structural routes (static
assets, health check, SSE). No path/method/response changes; verified via
an openapi()["paths"] diff (177 operations, byte-identical before/after)
and the full pytest suite (186 passed).
