+++ ev/interfaces/web/routes/notify.py
@@ -1,0 +1,44 @@
+"""Send-to-Telegram notify domain route — Phase 6b, Group 5 route split.
+Distinct from the notification-center domain (`notifications.py`, Group 1),
+which lists/reads/clears in-app notifications.
+
+Extract-and-recompose: logic moved verbatim from
+`ev/interfaces/web/app.py`'s `create_app()`.
+"""
+
+import asyncio
+
+from fastapi import APIRouter, Request
+
+from ..context import WebContext
+
+
+def build_router(ctx: WebContext) -> APIRouter:
+    router = APIRouter()
+    config = ctx.config
+
+    @router.post("/api/notify")
+    async def api_notify(request: Request):
+        """Send a message to the owner's own Telegram (a note to yourself)."""
+        ctx.check(request.headers.get("authorization"))
+        text = ((await ctx.body(request)).get("text") or "").strip()
+        if not text:
+            return {"ok": False, "msg": "Mensagem vazia."}
+        if not config.telegram_token or config.owner_id is None:
+            return {"ok": False, "msg": "Telegram não está configurado."}
+        import httpx
+
+        def _send():
+            return httpx.post(
+                f"https://api.telegram.org/bot{config.telegram_token}/sendMessage",
+                data={"chat_id": config.owner_id, "text": text}, timeout=15,
+            )
+        try:
+            r = await asyncio.to_thread(_send)
+            if r.status_code == 200:
+                return {"ok": True, "msg": "Mensagem enviada ao seu Telegram."}
+            return {"ok": False, "msg": f"Falha ao enviar (HTTP {r.status_code})."}
+        except Exception as exc:
+            return {"ok": False, "msg": f"Falha ao enviar: {exc}"}
+
+    return router
